+++ nikola/plugins/command/auto.plugin
